SpawnerAPI: Renamed the Client.lua file to SpawnerAPIClient.lua
