+++ media/lua/client/CommunityAPI.lua
@@ -4,7 +4,7 @@ CommunityAPI.Client = {
     ItemTooltip = require("ItemTooltipAPI/Client"),
     IsoRangeScan = require("IsoRangeScanAPI/Client"),
     Light = require("LightAPI/Client"),
-    Spawner = require("SpawnerAPI/Client"),
+    Spawner = require("SpawnerAPI/SpawnerAPIClient"),
     WorldSound = require("WorldSoundAPI/Client"),
 }
